+++ pdfmake-types.d.ts
@@ -1,8 +1,17 @@
-import { Style, TableOfContent, TableCellProperties } from 'pdfmake/interfaces'
+import {
+  Style,
+  TableOfContent,
+  TableCellProperties,
+  ContentColumns,
+} from 'pdfmake/interfaces'
 
 // Types updates not yet released to DefinitelyTyped
 declare module 'pdfmake/interfaces' {
   interface TableCellProperties {
     verticalAlignment?: 'top' | 'middle' | 'bottom' | undefined
+  }
+
+  interface ContentColumns {
+    snakingColumns?: boolean | undefined
   }
 }